refactor: move API Platform integration outside the Symfony bundle
