+++ src/ApiPlatform/Filters/NumericFilter.php
@@ -2,21 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Talleu\RedisOm\Bundle\ApiPlatform\Filters;
+namespace Talleu\RedisOm\ApiPlatform\Filters;
 
 use ApiPlatform\Metadata\ApiFilter;
 use ApiPlatform\Metadata\FilterInterface;
 use ApiPlatform\Metadata\Operation;
 
-class SearchFilter extends RedisAbstractFilter
+class NumericFilter extends RedisAbstractFilter
 {
     public function apply(array $params, string $resourceClass, ?Operation $operation = null, array $context = []): array
     {
         foreach ($context['filters'] as $property => $value) {
-            if ('order' === $property) {
-                continue;
-            }
-
             if (!property_exists($resourceClass, $property)) {
                 continue;
             }
@@ -25,16 +21,12 @@ class SearchFilter extends RedisAbstractFilter
                 continue;
             }
 
-            $strategy = $this->properties[$property];
-            if ('partial' === $strategy) {
-                $params['search_strategy'] = 'partial';
-            }
-
             $params['criteria'][$property] = $value;
         }
 
         return $params;
     }
+
 
     public function getDescription(string $resourceClass): array
     {
